refactor: progress, slow speed rerender
