+++ client/src/pages/contactList/index.js
@@ -471,7 +471,13 @@ const defaultState = {
 
 function ContactList () {
   const [state, setState] = useState(defaultState)
-
+  const sortedContacts = [...state.contacts].sort((a, b) => a.first_name < b.first_name ? -1 : 1)
+  useState(()=>{
+    setState(prev=>({
+      ...prev,
+      contacts:sortedContacts
+    }))
+  },[])
   const eventsHandler = () => {
 
   }
